added new classes AlphaBetaNode, CheckersHeuristic, and GameHeuristic.
AlphaBetaSearch not functioning yet, working on defining a game's state at a particular turn of the game.  If possible, I'd like for us to define the data structures and attributes that make up the action(s) made during a player's turn.  This could make it simpler to write code relating to move sequences, and we could add additional methods like iterators if we wanted.  Regarding the AlphaBetaNode class, it may allow us to simplify the algorithm in the AlphaBetaSearch class.  However, there may be a way to simplify the algorithm without adding additional classes, which is probably a good thing.  If anyone needs methods or attributes added to the interfaces/classes, please let me know :)