changed how Players are compared, almost done with AlphaBetaSearch
Unfortunately, cloning the CheckersBoard when running the algorithm in AlphaBetaSearch (ABS) causes comparisons between Players not to work right (The same Player will not be considered equivalent to itself in a cloned version).  So, I had to change some of the methods to use the id attribute.  The good news is that ABS is almost done, I just have to store the Piece and MoveType in the AlphaBetaNodes, and return the appropriate pair when the algorithm is finished